pr-style: finishing touches, XML code docs, guards
